Seperated the reuseable components in the form.jsx file and refactored the code and now we can use form.jsx wherever we need to validato validate the form
